requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
zipp==3.17.0
